Convergence test for mixed effects models

check_convergence() provides an alternative convergence test for merMod-objects.

check_convergence(x, tolerance = 0.001, ...)

Arguments

x
A merMod-object.

tolerance
Indicates up to which value the convergence result is accepted. The smaller tolerance is, the stricter the test will be.

...
Currently not used.

Value

TRUE if convergence is fine and FALSE if convergence is suspicious. Additionally, the convergence value is returned as attribute.

Details

check_convergence() provides an alternative convergence test for merMod-objects, as discussed here and suggested by Ben Bolker in this comment. Further details can be found in convergence.

Examples

if (require("lme4")) {
  data(cbpp)
  set.seed(1)
  cbpp$x <- rnorm(nrow(cbpp))
  cbpp$x2 <- runif(nrow(cbpp))

  model <- glmer(
    cbind(incidence, size - incidence) ~ period + x + x2 + (1 + x | herd),
    data = cbpp,
    family = binomial()
  )

  check_convergence(model)
}

#> Loading required package: lme4

#> Loading required package: Matrix

#> [1] TRUE
#> attr(,"gradient")
#> [1] 0.0002803075