Compare performance of different models

compare_performance() computes indices of model performance for different models at once and hence allows comparison of indices across models.

compare_performance(
  ...,
  metrics = "all",
  rank = FALSE,
  bayesfactor = TRUE,
  verbose = TRUE
)

Arguments

...
Multiple model objects (also of different classes).

metrics
Can be "all", "common" or a character vector of metrics to be computed. See related documentation of object's class for details.

rank
Logical, if TRUE, models are ranked according to "best overall model performance". See 'Details'.

bayesfactor
Logical, if TRUE, a Bayes factor for model comparisons is possibly returned. See 'Details'.

verbose
Toggle off warnings.

Value

A data frame (with one row per model) and one column per "index" (see metrics).

Details

Bayes factor for Model Comparison

If all models were fit from the same data, compare_performance() returns an additional column named BF, which shows the Bayes factor (see bayesfactor_models) for each model against the denominator model. The first model is used as denominator model, and its Bayes factor is set to NA to indicate the reference model.

Ranking Models

When rank = TRUE, a new column Performance_Score is returned. This score ranges from 0% to 100%, higher values indicating better model performance. Calculation is based on normalizing all indices (i.e. rescaling them to a range from 0 to 1), and taking the mean value of all indices for each model. This is a rather quick heuristic, but might be helpful as exploratory index.

In particular when models are of different types (e.g. mixed models, classical linear models, logistic regression, ...), not all indices will be computed for each model. In case where an index can't be calculated for a specific model type, this model gets an NA value. All indices that have any NAs are excluded from calculating the performance score.

There is a plot()-method for compare_performance(), which creates a "spiderweb" plot, where the different indices are normalized and larger values indicate better model performance. Hence, points closer to the center indicate worse fit indices (see online-documentation for more details).

Note

There is also a plot()-method implemented in the see-package.

Examples

if (require("lme4")) {
  m1 <- lm(mpg ~ wt + cyl, data = mtcars)
  m2 <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
  m3 <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
  compare_performance(m1, m2, m3)
}

#> Warning: When comparing models, please note that probably not all models were fit from same data.

#> # Comparison of Model Performance Indices
#> 
#> Model |    Type |    AIC |    BIC | RMSE |   R2 | R2_adjusted | R2_Tjur | LOGLOSS | SCORE_LOG | SCORE_SPHERICAL |  PCP | R2_conditional | R2_marginal |  ICC
#> ------------------------------------------------------------------------------------------------------------------------------------------------------------
#> m1    |      lm | 156.01 | 161.87 | 2.44 | 0.83 |        0.82 |         |         |           |                 |      |                |             |     
#> m2    |     glm |  31.30 |  35.70 | 0.89 |      |             |    0.48 |    0.40 |    -14.90 |            0.10 | 0.74 |                |             |     
#> m3    | lmerMod |  77.32 |  89.36 | 0.28 |      |             |         |         |           |                 |      |           0.97 |        0.10 | 0.97

data(iris)
lm1 <- lm(Sepal.Length ~ Species, data = iris)
lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
compare_performance(lm1, lm2, lm3)

#> # Comparison of Model Performance Indices
#> 
#> Model | Type |    AIC |    BIC |   R2 | R2_adjusted | RMSE |          BF
#> ------------------------------------------------------------------------
#> lm1   |   lm | 231.45 | 243.49 | 0.62 |        0.61 | 0.51 |            
#> lm2   |   lm | 106.23 | 121.29 | 0.84 |        0.83 | 0.33 | 3.44553e+26
#> lm3   |   lm | 106.77 | 127.84 | 0.84 |        0.83 | 0.33 | 1.29920e+25

compare_performance(lm1, lm2, lm3, rank = TRUE)

#> # Comparison of Model Performance Indices
#> 
#> Model | Type |    AIC |    BIC |   R2 | R2_adjusted | RMSE |          BF | Performance_Score
#> --------------------------------------------------------------------------------------------
#> lm2   |   lm | 106.23 | 121.29 | 0.84 |        0.83 | 0.33 | 3.44553e+26 |            99.25%
#> lm3   |   lm | 106.77 | 127.84 | 0.84 |        0.83 | 0.33 | 1.29920e+25 |            83.00%
#> lm1   |   lm | 231.45 | 243.49 | 0.62 |        0.61 | 0.51 |        1.00 |             0.00%
#> 
#> Model lm2 (of class lm) performed best with an overall performance score of 99.25%.
#>