display.performance_model {performance}	R Documentation

Print tables in different output formats

Description

Prints tables (i.e. data frame) in different output formats.

Usage

## S3 method for class 'performance_model'
display(object, format = "markdown", digits = 2, caption = NULL, ...)

## S3 method for class 'performance_model'
print(x, digits = 3, layout = "horizontal", ...)

## S3 method for class 'performance_model'
print_md(
  x,
  digits = 2,
  caption = "Indices of model performance",
  layout = "horizontal",
  ...
)

## S3 method for class 'compare_performance'
print_md(
  x,
  digits = 2,
  caption = "Comparison of Model Performance Indices",
  layout = "horizontal",
  ...
)

Arguments

object, x
An object returned by one of the package's function, for example model_performance(), compare_performance(), or check_itemscale().

format
String, indicating the output format. Can be "markdown" "html", or "tt". format = "tt" creates a tinytable object, which is either printed as markdown or HTML table, depending on the environment. See insight::export_table() for details.

digits
Number of decimal places.

caption
Table caption as string. If NULL, no table caption is printed.

...
Arguments passed to other methods, e.g. format() (and thereby to insight::format_table() or insight::export_table(). See related documentation for details on available arguments. For example, to control digits for information criteria like AIC or BIC, use ic_digits = <value>.

layout
Table layout (can be either "horizontal" or "vertical").

Details

display() is useful when the table-output from functions, which is usually printed as formatted text-table to console, should be formatted for pretty table-rendering in markdown documents, or if knitted from rmarkdown to PDF or Word files. See vignette for examples.

Value

A character vector. If format = "markdown", the return value will be a character vector in markdown-table format.

Global Options to Customize Output when Printing

easystats_display_format: options(easystats_display_format = <value>) will set the default format for the display() methods. Can be one of "markdown", "html", or "tt".

Examples

model <- lm(mpg ~ wt + cyl, data = mtcars)
mp <- model_performance(model)
display(mp)
display(mp, digits = 3, ic_digits = 4)

[Package performance version 0.16.0 Index]