Model Performance

See the documentation for your object's class:

Frequentist Regressions

Mixed models

Bayesian models

CFA / SEM lavaan models

compare_performance() computes indices of model performance for different models at once and hence allows comparison of indices across models.

compare_performance(..., metrics = "all", rank = FALSE, verbose = TRUE)

model_performance(model, ...)

performance(model, ...)

Arguments

...
Arguments passed to or from other methods, resp. for compare_performance(), one or multiple model objects (also of different classes).

metrics
Can be "all" or a character vector of metrics to be computed. See related documentation of object's class for details.

rank
Logical, if TRUE, models are ranked according to "best overall model performance". A new column Performance_Score is returned. This score ranges from 0% to 100%, higher values indicating better model performance. Calculation is based on normalizing all indices (i.e. rescaling them to a range from 0 to 1), and taking the mean value of all indices for each model. This is a rather quick heuristic, but might be helpful as exploratory index.

verbose
Toggle off warnings.

model
Statistical model.

Value

For model_performance(), a data frame (with one row) and one column per "index" (see metrics). For compare_performance(), the same data frame with one row per model.

Details

If all models were fit from the same data, compare_performance() returns an additional column named BF, which shows the Bayes factor (see bayesfactor_models) for each model against the denominator model. The first model is used as denominator model, and its Bayes factor is set to NA to indicate the reference model.

There is a plot()-method for compare_performance(), which creates a "spiderweb" plot, where the different indices are normalized and larger values indicate better model performance. Hence, points closer to the center indicate worse fit indices (see online-documentation for more details).

Examples

library(lme4)

m1 <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(m1)

#> # Indices of model performance
#> 
#>     AIC     BIC   R2 R2_adjusted  RMSE
#>  156.01 161.873 0.83       0.819 2.444

m2 <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
m3 <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
compare_performance(m1, m2, m3)

#> Warning: When comparing models, please note that probably not all models were fit from same data.

#> # Comparison of Model Performance Indices
#> 
#>  Model    Type     AIC     BIC  RMSE   R2 R2_adjusted R2_Tjur LOGLOSS SCORE_LOG
#>     m1      lm 156.010 161.873 2.444 0.83       0.819      NA      NA        NA
#>     m2     glm  31.298  35.695 0.889   NA          NA   0.478   0.395   -14.903
#>     m3 lmerMod  77.320  89.362 0.279   NA          NA      NA      NA        NA
#>  SCORE_SPHERICAL   PCP R2_conditional R2_marginal   ICC
#>               NA    NA             NA          NA    NA
#>            0.095 0.743             NA          NA    NA
#>               NA    NA          0.972       0.096 0.969

data(iris)
lm1 <- lm(Sepal.Length ~ Species, data = iris)
lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
compare_performance(lm1, lm2, lm3)

#> # Comparison of Model Performance Indices
#> 
#>  Model Type     AIC     BIC    R2 R2_adjusted  RMSE           BF
#>    lm1   lm 231.452 243.494 0.619       0.614 0.510           NA
#>    lm2   lm 106.233 121.286 0.837       0.833 0.333 3.445528e+26
#>    lm3   lm 106.767 127.842 0.840       0.835 0.330 1.299196e+25

compare_performance(lm1, lm2, lm3, rank = TRUE)

#> # Comparison of Model Performance Indices
#> 
#>  Model Type     AIC     BIC    R2 R2_adjusted  RMSE           BF
#>    lm2   lm 106.233 121.286 0.837       0.833 0.333 3.445528e+26
#>    lm3   lm 106.767 127.842 0.840       0.835 0.330 1.299196e+25
#>    lm1   lm 231.452 243.494 0.619       0.614 0.510           NA
#>  Performance_Score
#>             99.25%
#>             83.00%
#>              0.00%
#> 
#> Model lm2 (of class lm) performed best with an overall performance score of 99.25%.
#>