model_performance.fa {performance}	R Documentation

Performance of FA / PCA models

Description

Compute indices of model performance for models from the psych package, and for parameters::factor_analysis() and item_omega().

Usage

## S3 method for class 'fa'
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
A model object of class fa (e.g., from psych::fa()), principal (e.g., from psych::principal()), or from parameters::factor_analysis() or item_omega().

metrics
Can be "all" or a character vector of metrics to be computed (some of "Chi2", "Chi2_df", "df", "p_Chi2", "RMSR", "RMSR_corrected", "TLI", "RMSEA", and "BIC". For omega-models, can also include "R2" and "Correlation".

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Details

The RMSR_corrected metric is the Root Mean Square Residual corrected for degrees of freedom (i.e., dividing by degrees of freedom rather than the number of observations).

For omega-models, the columns R2 and Correlation are measures of factor score adequacy. R2 refers to the multiple R square of scores with factors, while Correlation indicates the correlation of scores with factors.

Value

A data frame (with one row) and one column per "index" (see metrics).

Examples

out <- psych::fa(psychTools::bfi[, 1:25], 5)
model_performance(out)

out <- item_omega(mtcars, n = 3)
model_performance(out)

[Package performance version 0.16.0 Index]