Performance of Regression Models

Compute indices of model performance for regression models.

# S3 method for lm
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
A model.

metrics
Can be "all", "common" or a character vector of metrics to be computed (some of c("AIC", "AICc", "BIC", "R2", "R2_adj", "RMSE", "SIGMA", "LOGLOSS", "PCP", "SCORE")). "common" will compute AIC, BIC, R2 and RMSE.

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Value

A data frame (with one row) and one column per "index" (see metrics).

Details

Depending on model, following indices are computed:

AIC Akaike's Information Criterion, see ?stats::AIC

AICc Second-order (or small sample) AIC with a correction for small sample sizes

BIC Bayesian Information Criterion, see ?stats::BIC

R2 r-squared value, see r2()

R2_adj adjusted r-squared, see r2()

RMSE root mean squared error, see performance_rmse()

SIGMA residual standard deviation, see insight::get_sigma()

LOGLOSS Log-loss, see performance_logloss()

SCORE_LOG score of logarithmic proper scoring rule, see performance_score()

SCORE_SPHERICAL score of spherical proper scoring rule, see performance_score()

PCP percentage of correct predictions, see performance_pcp()

Examples

model <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)
#> # Indices of model performance
#> 
#> AIC     |     BIC |    R2 | R2 (adj.) |  RMSE | Sigma
#> -----------------------------------------------------
#> 156.010 | 161.873 | 0.830 |     0.819 | 2.444 | 2.568

model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
model_performance(model)
#> # Indices of model performance
#> 
#> AIC    |    BIC | Tjur's R2 |  RMSE | Sigma | Log_loss | Score_log | Score_spherical |   PCP
#> --------------------------------------------------------------------------------------------
#> 31.298 | 35.695 |     0.478 | 0.359 | 0.934 |    0.395 |   -14.903 |           0.095 | 0.743