Performance of Meta-Analysis Models

Compute indices of model performance for meta-analysis model from the metafor package.

# S3 method for rma
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
A rma object as returned by metafor::rma().

metrics
Can be "all" or a character vector of metrics to be computed (some of c("AIC", "BIC", "I2", "H2", "TAU2", "R2", "CochransQ", "QE", "Omnibus", "QM")).

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Value

A data frame (with one row) and one column per "index" (see metrics).

Details

Indices of fit

AIC Akaike's Information Criterion, see AIC

BIC Bayesian Information Criterion, see BIC

I2: For a random effects model, I2 estimates (in percent) how much of the total variability in the effect size estimates can be attributed to heterogeneity among the true effects. For a mixed-effects model, I2 estimates how much of the unaccounted variability can be attributed to residual heterogeneity.

H2: For a random-effects model, H2 estimates the ratio of the total amount of variability in the effect size estimates to the amount of sampling variability. For a mixed-effects model, H2 estimates the ratio of the unaccounted variability in the effect size estimates to the amount of sampling variability.

TAU2: The amount of (residual) heterogeneity in the random or mixed effects model.

CochransQ (QE): Test for (residual) Heterogeneity. Without moderators in the model, this is simply Cochran's Q-test.

Omnibus (QM): Omnibus test of parameters.

R2: Pseudo-R2-statistic, which indicates the amount of heterogeneity accounted for by the moderators included in a fixed-effects model.

See the documentation for fitstats.

Examples

if (require("metafor")) {
  data(dat.bcg)
  dat <- escalc(measure = "RR", ai = tpos, bi = tneg, ci = cpos, di = cneg, data = dat.bcg)
  model <- rma(yi, vi, data = dat, method= "REML")
  model_performance(model)
}

#> Loading required package: metafor

#> Loading 'metafor' package (version 2.4-0). For an overview 
#> and introduction to the package please type: help(metafor).

#> # Indices of model performance
#> 
#>   AIC |   BIC |   I2 |    H2 | TAU2 | CochransQ | p_CochransQ | Omnibus | p_Omnibus
#> -----------------------------------------------------------------------------------
#> 28.41 | 29.38 | 0.92 | 12.86 | 0.31 |    152.23 |    p < .001 |   15.80 |  p < .001