Compute AIC and second order AIC

Compute the second-order Akaike's information criterion (AICc). The second-order (or small sample) is a AIC with a correction for small sample sizes. performance_aic() is a small wrapper that returns the AIC. It is a generic function that also works for some models that don't have a AIC method (like Tweedie models).

performance_aicc(x, ...)

performance_aic(x, ...)

Arguments

x
A model object.

...
Currently not used.

Value

Numeric, the AIC or AICc value.

References

Akaike, H. (1973) Information theory as an extension of the maximum likelihood principle. In: Second International Symposium on Information Theory, pp. 267–281. Petrov, B.N., Csaki, F., Eds, Akademiai Kiado, Budapest.

Hurvich, C. M., Tsai, C.-L. (1991) Bias of the corrected AIC criterion for underfitted regression and time series models. Biometrika 78, 499–509.

Examples

m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
AIC(m)

#> [1] 159.1051

performance_aicc(m)

#> [1] 162.4651