Likelihood-Ratio-Test for Model Comparison

Compute Likelihood-Ratio-Test for model comparison.

performance_lrt(...)

Arguments

...
Multiple model objects, which should respond to anova().

Value

A data frame, based on the results from anova().

Details

This only makes statistical sense if the models are nested. It is conventional to list the models from smallest to largest, but this is up to the user. The output shows the tests of the models against one another in the order specified.

See also

Examples

m1 <- lm(mpg ~ wt + cyl, data = mtcars)
m2 <- lm(mpg ~ wt + cyl + gear, data = mtcars)
m3 <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
performance_lrt(m1, m2, m3)

#> # Likelihood-Ratio-Test for Model Comparison
#> 
#> Model | Type | Df | AIC |     p
#> -------------------------------
#> m1    |   lm |  3 | 156 |      
#> m2    |   lm |  4 | 157 | 0.507
#> m3    |   lm |  5 | 159 | 0.563
#> 
#> Model 'm1' seems to have the best model fit.
#>