performance_rse {performance}	R Documentation

Residual Standard Error for Linear Models

Description

Compute residual standard error of linear models.

Usage


performance_rse(model)

Arguments

model
A model.

Details

The residual standard error is the square root of the residual sum of squares divided by the residual degrees of freedom.

Value

Numeric, the residual standard error of model.

Examples


data(mtcars)
m <- lm(mpg ~ hp + gear, data = mtcars)
performance_rse(m)

[Package performance version 0.11.0 Index]