Principal Components Analysis

This function performs a principal component analysis and returns the loadings (of the unrotated matrix) as data frame, or returns a rotated matrix of the loadings (if rotation is not NULL).

principal_components(x, rotation = NULL, n_comp = NULL)

Arguments

x
A data frame or a prcomp-object.

rotation
Rotation of the factor loadings. May be one of "varimax", "quartimax", "promax", "oblimin", "simplimax", "cluster" or "none". If rotatoin = NULL, loadings for the principal components from the unrotated matrix are returned.

n_comp
Number of components to extract. If rotation = "varmiax" and n_comp = NULL, number of components is based on the Kaiser-criteria.

Value

If rotation = NULL, a data frame with all loadings of principal components. Else, a rotated loadings matrix, as data frame. Details on the variance components are saved as attributes.

Details

The print()-method has a cutoff-argument, which is a scalar between 0 and 1, indicating which (absolute) values from the rotated loadings (i.e. when rotation is not NULL) should be blank in the output. By default, all loadings between -.1 and .1 are not shown.

Examples

data(iris)
principal_components(iris[, 1:4])

#>                       PC1   PC2   PC3   PC4
#> Standard deviation  1.708 0.956 0.383 0.144
#> Eigenvalue          2.918 0.914 0.147 0.021
#> Proportion variance 0.730 0.229 0.037 0.005
#> Cumulative variance 0.730 0.958 0.995 1.000

data(iris)
principal_components(iris[, 1:4], rotation = "varimax", n_comp = 2)

#> # Rotated loadings from Principal Component Analysis (varimax-rotation)
#> 
#>                 PC1    PC2
#> Sepal.Length  0.959       
#> Sepal.Width  -0.145 -0.985
#> Petal.Length  0.944  0.304
#> Petal.Width   0.932  0.257
#> 
#> (Explained) Variance
#>                        PC1   PC2
#> Proportion variance  0.676 0.283
#> Cumulative variance  0.676 0.958
#> Proportion explained 0.705 0.295
#> Cumulative explained 0.705 1.000

pr <- principal_components(iris[, 1:4], rotation = "varimax", n_comp = 2)

# show all
print(pr, cutoff = .001)

#> # Rotated loadings from Principal Component Analysis (varimax-rotation)
#> 
#>                 PC1    PC2
#> Sepal.Length  0.959 -0.048
#> Sepal.Width  -0.145 -0.985
#> Petal.Length  0.944  0.304
#> Petal.Width   0.932  0.257
#> 
#> (Explained) Variance
#>                        PC1   PC2
#> Proportion variance  0.676 0.283
#> Cumulative variance  0.676 0.958
#> Proportion explained 0.705 0.295
#> Cumulative explained 0.705 1.000

# show only some
print(pr, cutoff = .5)

#> # Rotated loadings from Principal Component Analysis (varimax-rotation)
#> 
#>                PC1    PC2
#> Sepal.Length 0.959       
#> Sepal.Width        -0.985
#> Petal.Length 0.944       
#> Petal.Width  0.932       
#> 
#> (Explained) Variance
#>                        PC1   PC2
#> Proportion variance  0.676 0.283
#> Cumulative variance  0.676 0.958
#> Proportion explained 0.705 0.295
#> Cumulative explained 0.705 1.000