Likelihood-Ratio-Test for Model Comparison

Compute Likelihood-Ratio-Test (LRT) for model comparison, which tests which model is a better (more likely) explanation of the data. Likelihood-Ratio-Test (LRT) gives usually somewhat close results (if not equivalent) to the commonly used Wald F-test (obtained by default with anova(...)). Maximum likelihood tests make stronger assumptions than method of moments tests like the F-test, and in turn are more efficient. Agresti (1990) suggests that you should use the LRT instead of the Wald test for small sample sizes or if the parameters are large. A "small" sample size is under about 30.

For regression models, this is similar to anova(..., test="LRT") or lmtest::lrtest(...), depending on the estimator.

For lavaan models (SEM, CFA), the function calls lavaan::lavTestLRT().

test_lrt(...)

test_likelihoodratio(...)

performance_lrt(...)

# S3 method for default
test_lrt(..., estimator = "ML")

Arguments

...
Multiple model objects.

estimator
Applied when comparing regression models. Corresponds to the different estimators for the standard deviation of the errors. If estimator="OLS" (default), then it uses the same method as anova(..., test="LRT") implemented in base R, i.e., scaling by n-k (the unbiased OLS estimator) and using this estimator under the alternative hypothesis. If estimator="ML", which is for instance used by lrtest(...) in package lmtest, the scaling is done by n (the biased ML estimator) and the estimator under the null hypothesis. In moderately large samples, the differences should be negligible, but it is possible that OLS would perform slightly better in small samples with Gaussian errors.

Value

A data frame, based on the results from anova().

Details

This only makes statistical sense if the models are nested. It is conventional to list the models from smallest to largest, but this is up to the user. The output shows the tests of the models against one another in the order specified.

See also

Examples

# Regression Models
m1 <- lm(mpg ~ wt + cyl, data = mtcars)
m2 <- lm(mpg ~ wt + cyl + gear, data = mtcars)
m3 <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)

test_lrt(m1, m2, m3, estimator = "ML") # Equivalent to lmtest::lrtest(m1, m2, m3)

#> # Likelihood-Ratio-Test (LRT) for Model Comparison
#> 
#> Model | df | df_diff | Chi2 |     p
#> -----------------------------------
#> m1    |  4 |         |      |      
#> m2    |  5 |       1 | 0.51 | 0.475
#> m3    |  6 |       1 | 0.39 | 0.530

test_lrt(m1, m2, m3, estimator = "OLS") # Equivalent to anova(m1, m2, m3, test='LRT')

#> # Likelihood-Ratio-Test (LRT) for Model Comparison
#> 
#> Model | df | df_diff | Chi2 |     p
#> -----------------------------------
#> m1    |  4 |         |      |      
#> m2    |  5 |       1 | 0.44 | 0.507
#> m3    |  6 |       1 | 0.33 | 0.563

# nested from "largest" to "smallest" model.
test_lrt(m3, m2, m1)

#> # Likelihood-Ratio-Test (LRT) for Model Comparison
#> 
#> Model | df | df_diff | Chi2 |     p
#> -----------------------------------
#> m3    |  6 |         |      |      
#> m2    |  5 |      -1 | 0.39 | 0.530
#> m1    |  4 |      -1 | 0.51 | 0.475

# Lavaan Models
if (require("lavaan")) {
  structure <- " visual  =~ x1 + x2 + x3
                 textual =~ x4 + x5 + x6
                 speed   =~ x7 + x8 + x9

                  visual ~~ textual + speed "
  m1 <- lavaan::cfa(structure, data = HolzingerSwineford1939)

  structure <- " visual  =~ x1 + x2 + x3
                 textual =~ x4 + x5 + x6
                 speed   =~ x7 + x8 + x9

                  visual ~~ 0 * textual + speed "
  m2 <- lavaan::cfa(structure, data = HolzingerSwineford1939)

  structure <- " visual  =~ x1 + x2 + x3
                 textual =~ x4 + x5 + x6
                 speed   =~ x7 + x8 + x9

                  visual ~~ 0 * textual + 0 * speed "
  m3 <- lavaan::cfa(structure, data = HolzingerSwineford1939)

  test_lrt(m1, m2, m3)
}

#> # Likelihood-Ratio-Test (LRT) for Model Comparison
#> 
#> Model |   Type | df | df_diff |   Chi2 |      p
#> -----------------------------------------------
#> m1    | lavaan | 24 |         |  85.31 |       
#> m2    | lavaan | 25 |       1 | 121.47 | < .001
#> m3    | lavaan | 26 |       1 | 140.60 | < .001