Test if Models are Different using Vuong's (1989) Test

Test if models are different using Vuong's (1989) test, which can be used both for nested and non-nested models. It actually consists of two tests. The test of distinguishability (the Omega2) indicates whether or not the models can possibly be distinguished on the basis of the observed data. If its p-value is significant, it means the models are distinguishable. The second test indicates whether each model fits better than the reference model. If the models are nested, then the test works as a robust LRT.

test_vuong(...)

Arguments

...
Multiple model objects.

Value

A data frame.

Details

The code for this function is adapted from the nonnest2 package, and all credit go to their authors.

References

Vuong, Q. H. (1989). Likelihood ratio tests for model selection and non-nested hypotheses. Econometrica, 57, 307-333.

Merkle, E. C., You, D., & Preacher, K. (2016). Testing non-nested structural equation models. Psychological Methods, 21, 151-163.

See also

Examples

# Nested Models
m1 <- lm(Sepal.Length ~ Petal.Width * Species, data = iris)
m2 <- lm(Sepal.Length ~ Petal.Width + Species, data = iris)
m3 <- lm(Sepal.Length ~ Petal.Width, data = iris)

test_vuong(m1, m2, m3) # nonnest2::vuongtest(m1, m2)

#>   Name Model       Omega2   p_Omega2        LR       p_LR
#> 1   m1    lm           NA         NA        NA         NA
#> 2   m2    lm 0.0225040529 0.08112211 3.4092877 0.09903392
#> 3   m3    lm 0.0009541423 0.93491015 0.1535725 0.91912847

# Non-nested
m1 <- lm(Sepal.Length ~ Petal.Width, data = iris)
m2 <- lm(Sepal.Length ~ Petal.Length, data = iris)
m3 <- lm(Sepal.Length ~ Species, data = iris)

test_vuong(m1, m2, m3) # nonnest2::vuongtest(m1, m2, nested=TRUE)

#>   Name Model    Omega2      p_Omega2        LR         p_LR
#> 1   m1    lm        NA            NA        NA           NA
#> 2   m2    lm 0.1855419 -2.242953e-06 -4.566458 2.480171e-06
#> 3   m3    lm 0.1193515  1.379073e-04  2.508828 6.056629e-03