Describe Properties of Item Scales

Compute various measures of internal consistencies applied to (sub)scales, which items were extracted using parameters::principal_components().

check_itemscale(x)

Arguments

x
An object of class parameters_pca, as returned by parameters::principal_components().

Value

A list of data frames, with related measures of internal consistencies of each subscale.

Details

check_itemscale() calculates various measures of internal consistencies, such as Cronbach's alpha, item difficulty or discrimination etc. on subscales which were built from several items. Subscales are retrieved from the results of parameters::principal_components(), i.e. based on how many components were extracted from the PCA, check_itemscale() retrieves those variables that belong to a component and calculates the above mentioned measures.

Note

Item difficulty should range between 0.2 and 0.8. Ideal value is p+(1-p)/2 (which mostly is between 0.5 and 0.8). See item_difficulty for details.

For item discrimination, acceptable values are 0.20 or higher; the closer to 1.00 the better. See item_reliability for more details.

In case the total Cronbach's alpha value is below the acceptable cut-off of 0.7 (mostly if an index has few items), the mean inter-item-correlation is an alternative measure to indicate acceptability. Satisfactory range lies between 0.2 and 0.4. See also item_intercor.

References

Briggs SR, Cheek JM (1986) The role of factor analysis in the development and evaluation of personality scales. Journal of Personality, 54(1), 106-148. doi: 10.1111/j.1467-6494.1986.tb00391.x

Trochim WMK (2008) Types of Reliability. (web)

Examples

# data generation from '?prcomp', slightly modified
C <- chol(S <- toeplitz(.9^(0:15)))
set.seed(17)
X <- matrix(rnorm(16000), 100, 16)
Z <- X %*% C
if (require("parameters") && require("psych")) {
  pca <- principal_components(as.data.frame(Z), rotation = "varimax", n = 3)
  pca
  check_itemscale(pca)
}

#> Loading required package: psych

#> # Description of (Sub-)Scales
#> Component 1
#> 
#> Item | Missings |  Mean |   SD | Skewness | Difficulty | Discrimination | alpha if deleted
#> ------------------------------------------------------------------------------------------
#> V1   |        0 | -0.02 | 1.06 |    -0.49 |      -0.01 |           0.80 |             0.96
#> V2   |        0 | -0.05 | 1.05 |    -0.29 |      -0.02 |           0.90 |             0.95
#> V3   |        0 |  0.00 | 1.10 |    -0.77 |       0.00 |           0.94 |             0.95
#> V4   |        0 |  0.00 | 1.10 |    -0.82 |       0.00 |           0.92 |             0.95
#> V5   |        0 | -0.07 | 1.09 |    -0.29 |      -0.03 |           0.90 |             0.95
#> V6   |        0 | -0.04 | 1.13 |    -0.27 |      -0.01 |           0.83 |             0.96
#> 
#> Mean inter-item-correlation = 0.813  Cronbach's alpha = 0.963
#> 
#> Component 2
#> 
#> Item | Missings |  Mean |   SD | Skewness | Difficulty | Discrimination | alpha if deleted
#> ------------------------------------------------------------------------------------------
#> V7   |        0 | -0.01 | 1.07 |     0.01 |       0.00 |           0.87 |             0.97
#> V8   |        0 |  0.02 | 0.96 |     0.23 |       0.01 |           0.89 |             0.96
#> V9   |        0 |  0.04 | 0.98 |     0.37 |       0.01 |           0.93 |             0.96
#> V10  |        0 |  0.08 | 1.00 |     0.18 |       0.03 |           0.93 |             0.96
#> V11  |        0 |  0.02 | 1.03 |     0.18 |       0.01 |           0.92 |             0.96
#> V12  |        0 |  0.00 | 1.04 |     0.27 |       0.00 |           0.84 |             0.97
#> 
#> Mean inter-item-correlation = 0.840  Cronbach's alpha = 0.969
#> 
#> Component 3
#> 
#> Item | Missings |  Mean |   SD | Skewness | Difficulty | Discrimination | alpha if deleted
#> ------------------------------------------------------------------------------------------
#> V13  |        0 |  0.04 | 0.95 |     0.10 |       0.02 |           0.81 |             0.95
#> V14  |        0 | -0.02 | 0.96 |     0.24 |      -0.01 |           0.93 |             0.91
#> V15  |        0 | -0.03 | 0.94 |     0.41 |      -0.01 |           0.92 |             0.91
#> V16  |        0 |  0.03 | 0.96 |     0.28 |       0.01 |           0.82 |             0.94
#> 
#> Mean inter-item-correlation = 0.811  Cronbach's alpha = 0.945