Check for zero-inflation in count models

check_zeroinflation() checks whether count models are over- or underfitting zeros in the outcome.

check_zeroinflation(x, tolerance = 0.05)

Arguments

x

Fitted model of class merMod, glmmTMB, glm, or glm.nb (package MASS).

tolerance

The tolerance for the ratio of observed and predicted zeros to considered as over- or underfitting zeros. A ratio between 1 +/- tolerance is considered as OK, while a ratio beyond or below this threshold would indicate over- or underfitting.

Value

A list with information about the amount of predicted and observed zeros in the outcome, as well as the ratio between these two values.

Details

If the amount of observed zeros is larger than the amount of predicted zeros, the model is underfitting zeros, which indicates a zero-inflation in the data. In such cases, it is recommended to use negative binomial or zero-inflated models.

Examples

if (require("glmmTMB")) {
  data(Salamanders)
  m <- glm(count ~ spp + mined, family = poisson, data = Salamanders)
  check_zeroinflation(m)
}
#> # Check for zero-inflation
#> 
#>    Observed zeros: 387
#>   Predicted zeros: 298
#>             Ratio: 0.77
#> 
#> Model is underfitting zeros (probable zero-inflation).