Dominance Analysis

Computes Dominance Analysis Statistics and Designations

Usage

dominance_analysis(
  model,
  sets = NULL,
  all = NULL,
  conditional = TRUE,
  complete = TRUE,
  ...
)

Arguments

model

A model object supported by performance::r2(). See 'Details'.

sets

A (named) list of formula objects with no left hand side/response. A named list uses the name provided each element as the label for the set.

Predictors in each formula are bound together as a set in the dominance analysis and dominance statistics and designations are computed for the predictors together. Predictors in sets must be present in the model submitted to the model argument and cannot be in the all argument.

all

A formula with no left hand side/response.

Predictors in the formula are included in each subset in the dominance analysis and the R2 value associated with them is subtracted from the overall value. Predictors in all must be present in the model submitted to the model argument and cannot be in the sets argument.

conditional

Logical. If FALSE then conditional dominance matrix is not computed.

If conditional dominance is not desired as an importance criterion, avoiding computing the conditional dominance matrix can save computation time.

complete

Logical. If FALSE then complete dominance matrix is not computed.

If complete dominance is not desired as an importance criterion, avoiding computing complete dominance designations can save computation time.

...

Not used at current.

Value

Object of class "dominance_analysis". An object of class "dominance_analysis" is a list composed of the following elements:

general_dominance

Vector of general dominance statistics.

standardized

Vector of general dominance statistics normalized to sum to 1.

ranks

Vector of ranks applied to the general dominance statistics.

conditional_dominance

Matrix of conditional dominance statistics. Each row represents a predictor; each column represents an the average increment to R2 with a specific number of predictors in the model or NULL.

complete_dominance

Logical matrix of complete dominance designations. The predictors represented in each row are cross-referenced with predictors in each column. Whether the predictor in each column dominates the predictor in each row is represented in the logical value in the matrix or NULL.

model_R2

Value of R2 value returned by the r2() method for the model.

all_subset_R2

Value of R2 associated with the predictors in the all argument or NULL.

Details

Computes two decompositions of the model's R2 and returns a matrix of designations from which predictor relative importance determinations can be obtained.

The input model is parsed using insight::find_predictors(), does not yet support interactions, transformations, or offsets applied in the R formula, and will fail with an error if any such terms are detected.

The model submitted must accept an formula object as a formula argument. In addition, the model object must accept the data on which the model is estimated as a data argument. Formulas submitted using object references (i.e., lm(mtcars$mpg ~ mtcars$vs)) and functions that accept data as a non-data argument (e.g., survey::svyglm() uses design) will fail.

Models that return TRUE for the insight::model_info() function's values "is_bayesian", "is_mixed", "is_gam", is_multivariate", "is_zero_inflated", or "is_hurdle" are not supported at current.

When performance::r2() returns multiple values, only the first is used by default.

The underlying domir::domin() function that implements the dominance statistic and designation computations has only been tested to R version 3.5 and will fail with an error if called in R versions < 3.5.

References

Azen, R., & Budescu, D. V. (2003). The dominance analysis approach for comparing predictors in multiple regression. Psychological Methods, 8(2), 129-148. doi:10.1037/1082-989X.8.2.129

Budescu, D. V. (1993). Dominance analysis: A new approach to the problem of relative importance of predictors in multiple regression. Psychological Bulletin, 114(3), 542-551. doi:10.1037/0033-2909.114.3.542

Groemping, U. (2007). Estimators of relative importance in linear regression based on variance decomposition. The American Statistician, 61(2), 139-147. doi:10.1198/000313007X188252

Luchman, J. N., Lei, X., & Kaplan, S. A. (2020). Relative Importance Analysis With Multivariate Models: Shifting the Focus from Independent Variables to Parameter Estimates. Journal of Applied Structural Equation Modeling, 4(2), 1-20. doi:10.47263/JASEM.4(2)02

Luchman, J. N. (2021). Determining relative importance in Stata using dominance analysis: domin and domme. Stata Journal 21(2), 510-538. doi:10.1177/1536867X211025837

See also

Author

Joseph Luchman

Examples

if (getRversion() >= "3.5.0" && require("domir")) {
  data(mtcars)
  model <- glm(vs ~ cyl + carb + mpg, data = mtcars, family = binomial())

  r2(model)
  dominance_analysis(model)
}
#> Loading required package: domir
#>     parameter subset
#> 1 (Intercept)   <NA>
#> 2         cyl   <NA>
#> 3        carb   <NA>
#> 4         mpg   <NA>
#>     parameter   subset
#> 1 (Intercept) constant
#> 2         cyl      cyl
#> 3        carb     carb
#> 4         mpg      mpg
#> # Dominance Analysis Results
#> 
#> Model R2 Value:  0.741 
#> 
#> parameter   | general_dominance | standardized | ranks |   subset
#> -----------------------------------------------------------------
#> (Intercept) |                   |              |       | constant
#> cyl         |             0.366 |        0.493 |     1 |      cyl
#> carb        |             0.178 |        0.241 |     3 |     carb
#> mpg         |             0.197 |        0.266 |     2 |      mpg
#> 
#> subset | IVs_1 | IVs_2 | IVs_3
#> ------------------------------
#> cyl    | 0.654 | 0.254 | 0.190
#> carb   | 0.394 | 0.066 | 0.074
#> mpg    | 0.474 | 0.085 | 0.032
#> 
#> subset | Dmnated_cyl | Dmnated_carb | Dmnated_mpg
#> -------------------------------------------------
#> cyl    |             |        FALSE |       FALSE
#> carb   |        TRUE |              |            
#> mpg    |        TRUE |              |