Difficulty of Questionnaire Items

Compute various measures of internal consistencies for tests or item-scales of questionnaires.

item_difficulty(x)

Arguments

x
Depending on the function, x may be a matrix as returned by the cor()-function, or a data frame with items (e.g. from a test or questionnaire).

Value

A data frame with three columns: The name(s) of the item(s), the item difficulties for each item, and the ideal item difficulty.

Details

This function calculates the item difficulty, which should range between 0.2 and 0.8. Lower values are a signal for more difficult items, while higher values close to one are a sign for easier items. The ideal value for item difficulty is p + (1 - p) / 2, where p = 1 / max(x). In most cases, the ideal item difficulty lies between 0.5 and 0.8.

Examples

data(mtcars)
x <- mtcars[, c("cyl", "gear", "carb", "hp")]
item_difficulty(x)
#> # Item Difficulty
#> 
#>       difficulty  ideal
#>    cyl      0.77   0.56
#>   gear      0.74   0.60
#>   carb      0.35   0.56
#>     hp      0.44   0.50