item_discrimination {performance}	R Documentation

Discrimination of Questionnaire Items

Description

Compute various measures of internal consistencies for tests or item-scales of questionnaires.

Usage


item_discrimination(x, standardize = FALSE)

Arguments

x
A matrix or a data frame.

standardize
Logical, if TRUE, the data frame's vectors will be standardized. Recommended when the variables have different measures / scales.

Details

This function calculates the item discriminations (corrected item-total correlations for each item of x with the remaining items) for each item of a scale. The absolute value of the item discrimination indices should be above 0.2. An index between 0.2 and 0.4 is considered as "fair", while a satisfactory index ranges from 0.4 to 0.7. Items with low discrimination indices are often ambiguously worded and should be examined. Items with negative indices should be examined to determine why a negative value was obtained (e.g. reversed answer categories regarding positive and negative poles).

Value

A data frame with the item discrimination (corrected item-total correlations) for each item of the scale.

References

Kelava A, Moosbrugger H (2020). Deskriptivstatistische Itemanalyse und Testwertbestimmung. In: Moosbrugger H, Kelava A, editors. Testtheorie und Fragebogenkonstruktion. Berlin, Heidelberg: Springer, 143–158

Examples


data(mtcars)
x <- mtcars[, c("cyl", "gear", "carb", "hp")]
item_discrimination(x)

[Package performance version 0.11.0 Index]