Reliability Test for Items or Scales

Compute various measures of internal consistencies for tests or item-scales of questionnaires.

item_reliability(x, standardize = FALSE, digits = 3)

Arguments

x
A matrix or a data frame.

standardize
Logical, if TRUE, the data frame's vectors will be standardized. Recommended when the variables have different measures / scales.

digits
Amount of digits for returned values.

Value

A data frame with the corrected item-total correlations (item discrimination, column item_discrimination) and Cronbach's Alpha (if item deleted, column alpha_if_deleted) for each item of the scale, or NULL if data frame had too less columns.

Details

This function calculates the item discriminations (corrected item-total correlations for each item of x with the remaining items) and the Cronbach's alpha for each item, if it was deleted from the scale. The absolute value of the item discrimination indices should be above 0.1. An index between 0.1 and 0.3 is considered as "fair", while an index above 0.3 (or below -0.3) is "good". Items with low discrimination indices are often ambiguously worded and should be examined. Items with negative indices should be examined to determine why a negative value was obtained (e.g. reversed answer categories regarding positive and negative poles).

Examples

data(mtcars)
x <- mtcars[, c("cyl", "gear", "carb", "hp")]
item_reliability(x)
#>   term alpha_if_deleted item_discrimination
#> 1  cyl            0.048               0.826
#> 2 gear            0.110              -0.127
#> 3 carb            0.058               0.751
#> 4   hp            0.411               0.881