LOO-related Indices for Bayesian regressions.

Compute LOOIC (leave-one-out cross-validation (LOO) information criterion) and ELPD (expected log predictive density) for Bayesian regressions.

looic(model, verbose = TRUE)

Arguments

model
A Bayesian regression model.

verbose
Toggle off warnings.

Value

A list with four elements, the ELPD, LOOIC and their standard errors.

Examples

if (require("rstanarm")) {
  model <- stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
  looic(model)
}

#> Loading required package: rstanarm

#> Loading required package: Rcpp

#> This is rstanarm version 2.21.1

#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!

#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#>   options(mc.cores = parallel::detectCores())

#> 
#> Attaching package: ‘rstanarm’

#> The following object is masked from ‘package:psych’:
#> 
#>     logit

#> The following object is masked from ‘package:performance’:
#> 
#>     pp_check

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> # LOOIC and ELPD with Standard Error
#> 
#>   LOOIC: 155.90 [8.79]
#>    ELPD: -77.95 [4.39]