Model Performance

See the documentation for your object's class: (generalized) linear models, mixed models and Bayesian models. compare_performance() computes indices of model performance for different models at once and hence allows comparison of indices accros models.

model_performance(model, ...)

compare_performance(..., metrics = "all")

Arguments

model
Statistical model.

...
Arguments passed to or from other methods, resp. for compare_performance(), one or multiple model objects (also of different classes).

metrics
Can be "all" or a character vector of metrics to be computed. See related documentation of object's class for details.

Value

For model_performance(), a data frame (with one row) and one column per "index" (see metrics). For compare_performance(), the same data frame with one row per model.

Examples

library(lme4)

m1 <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(m1)

#>        AIC     BIC        R2 R2_adjusted     RMSE
#> 1 156.0101 161.873 0.8302274   0.8185189 2.444202

m2 <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
m3 <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
compare_performance(m1, m2, m3)

#>   name   class       AIC       BIC      RMSE        R2 R2_adjusted   R2_Tjur
#> 1   m1     glm  31.29788  35.69508 0.8891336        NA          NA 0.4776926
#> 2   m2      lm 156.01007 161.87301 2.4442021 0.8302274   0.8185189        NA
#> 3   m3 lmerMod  77.31960  89.36214 0.2787760        NA          NA        NA
#>     LOGLOSS       PCP R2_conditional R2_marginal       ICC
#> 1 0.3952793 0.7429268             NA          NA        NA
#> 2        NA        NA             NA          NA        NA
#> 3        NA        NA      0.9723224   0.0955834 0.9693973