Model Performance

See the documentation for your object's class:

Frequentist Regressions

Instrumental Variables Regressions

Mixed models

Bayesian models

CFA / SEM lavaan models

Meta-analysis models

Usage

model_performance(model, ...)

performance(model, ...)

Arguments

model

Statistical model.

...

Arguments passed to or from other methods, resp. for compare_performance(), one or multiple model objects (also of different classes).

Value

A data frame (with one row) and one column per "index" (see metrics).

Details

model_performance() correctly detects transformed response and returns the "corrected" AIC and BIC value on the original scale. To get back to the original scale, the likelihood of the model is multiplied by the Jacobian/derivative of the transformation.

See also

Examples

model <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)
#> # Indices of model performance
#> 
#> AIC     |    AICc |     BIC |    R2 | R2 (adj.) |  RMSE | Sigma
#> ---------------------------------------------------------------
#> 156.010 | 157.492 | 161.873 | 0.830 |     0.819 | 2.444 | 2.568

model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
model_performance(model)
#> # Indices of model performance
#> 
#> AIC    |   AICc |    BIC | Tjur's R2 |  RMSE | Sigma | Log_loss | Score_log
#> ---------------------------------------------------------------------------
#> 31.298 | 32.155 | 35.695 |     0.478 | 0.359 | 1.000 |    0.395 |   -14.903
#> 
#> AIC    | Score_spherical |   PCP
#> --------------------------------
#> 31.298 |           0.095 | 0.743