model_performance {performance}	R Documentation

Model Performance

Description

See the documentation for your object's class:

Frequentist Regressions

Instrumental Variables Regressions

Mixed models

Bayesian models

CFA / SEM lavaan models

Meta-analysis models

Usage


model_performance(model, ...)

performance(model, ...)

Arguments

model
Statistical model.

...
Arguments passed to or from other methods, resp. for compare_performance(), one or multiple model objects (also of different classes).

Details

model_performance() correctly detects transformed response and returns the "corrected" AIC and BIC value on the original scale. To get back to the original scale, the likelihood of the model is multiplied by the Jacobian/derivative of the transformation.

Value

A data frame (with one row) and one column per "index" (see metrics).

See Also

compare_performance() to compare performance of many different models.

Examples


model <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)

model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
model_performance(model)

[Package performance version 0.11.0 Index]