Performance of Regression Models

Compute indices of model performance for regression models.

# S3 method for lm
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
A model.

metrics
Can be "all", "common" or a character vector of metrics to be computed (some of c("AIC", "BIC", "R2", "RMSE", "LOGLOSS", "PCP", "SCORE")). "common" will compute AIC, BIC, R2 and RMSE.

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Value

A data frame (with one row) and one column per "index" (see metrics).

Details

Depending on model, following indices are computed:

AIC Akaike's Information Criterion, see ?stats::AIC

BIC Bayesian Information Criterion, see ?stats::BIC

R2 r-squared value, see r2

R2_adj adjusted r-squared, see r2

RMSE root mean squared error, see performance_rmse

LOGLOSS Log-loss, see performance_logloss

SCORE_LOG score of logarithmic proper scoring rule, see performance_score

SCORE_SPHERICAL score of spherical proper scoring rule, see performance_score

PCP percentage of correct predictions, see performance_pcp

Examples

model <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)

#> # Indices of model performance
#> 
#>    AIC |    BIC |   R2 | R2_adjusted | RMSE
#> -------------------------------------------
#> 156.01 | 161.87 | 0.83 |        0.82 | 2.44

model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
model_performance(model)

#> # Indices of model performance
#> 
#>   AIC |   BIC | R2_Tjur | RMSE | LOGLOSS | SCORE_LOG | SCORE_SPHERICAL |  PCP
#> -----------------------------------------------------------------------------
#> 31.30 | 35.70 |    0.48 | 0.89 |    0.40 |    -14.90 |            0.10 | 0.74