model_performance.lm {performance}	R Documentation

Performance of Regression Models

Description

Compute indices of model performance for regression models.

Usage


## S3 method for class 'lm'
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
A model.

metrics
Can be "all", "common" or a character vector of metrics to be computed (one or more of "AIC", "AICc", "BIC", "R2", "R2_adj", "RMSE", "SIGMA", "LOGLOSS", "PCP", "SCORE"). "common" will compute AIC, BIC, R2 and RMSE.

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Details

Depending on model, following indices are computed:

AIC: Akaike's Information Criterion, see ?stats::AIC

AICc: Second-order (or small sample) AIC with a correction for small sample sizes

BIC: Bayesian Information Criterion, see ?stats::BIC

R2: r-squared value, see r2()

R2_adj: adjusted r-squared, see r2()

RMSE: root mean squared error, see performance_rmse()

SIGMA: residual standard deviation, see insight::get_sigma()

LOGLOSS: Log-loss, see performance_logloss()

SCORE_LOG: score of logarithmic proper scoring rule, see performance_score()

SCORE_SPHERICAL: score of spherical proper scoring rule, see performance_score()

PCP: percentage of correct predictions, see performance_pcp()

model_performance() correctly detects transformed response and returns the "corrected" AIC and BIC value on the original scale. To get back to the original scale, the likelihood of the model is multiplied by the Jacobian/derivative of the transformation.

Value

A data frame (with one row) and one column per "index" (see metrics).

Examples


model <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)

model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
model_performance(model)

[Package performance version 0.11.0 Index]