Performance of Mixed Models

Compute indices of model performance for mixed models.

# S3 method for merMod
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
A mixed effects model.

metrics
Can be "all", "common" or a character vector of metrics to be computed (some of c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS", "SCORE")). "common" will compute AIC, BIC, R2, ICC and RMSE.

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Value

A data frame (with one row) and one column per "index" (see metrics).

Details

This method returns the adjusted ICC only, as this is typically of interest when judging the variance attributed to the random effects part of the model (see also icc).

Furthermore, see 'Details' in model_performance.lm for more details on returned indices.

Examples

library(lme4)
model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
model_performance(model)

#>       AIC      BIC R2_conditional R2_marginal       ICC     RMSE
#> 1 77.3196 89.36214      0.9723224   0.0955834 0.9693973 0.278776