Performance of Bayesian Models

Compute indices of model performance for (general) linear models.

# S3 method for stanreg
model_performance(model, metrics = "all", verbose = TRUE, ...)

Arguments

model
Object of class stanreg or brmsfit.

metrics
Can be "all", "common" or a character vector of metrics to be computed (some of c("LOOIC", "WAIC", "R2", "R2_adj", "RMSE", "LOGLOSS", "SCORE")). "common" will compute LOOIC, WAIC, R2 and RMSE.

verbose
Toggle off warnings.

...
Arguments passed to or from other methods.

Value

A data frame (with one row) and one column per "index" (see metrics).

Details

Depending on model, following indices are computed:

ELPD expected log predictive density, see looic

LOOIC leave-one-out cross-validation (LOO) information criterion, see looic

WAIC widely applicable information criterion, see loo::waic

R2 r-squared value, see r2

R2_LOO_adjusted adjusted r-squared, see r2

RMSE root mean squared error, see performance_rmse

LOGLOSS Log-loss, see performance_logloss

SCORE_LOG score of logarithmic proper scoring rule, see performance_score

SCORE_SPHERICAL score of spherical proper scoring rule, see performance_score

PCP percentage of correct predictions, see performance_pcp

References

Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, The American Statistician, 1-6.

See also

Examples

if (require("rstanarm")) {
  model <- stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
  model_performance(model)

  model <- stan_glmer(
    mpg ~ wt + cyl + (1 | gear),
    data = mtcars,
    chains = 1,
    iter = 500,
    refresh = 0
  )
  model_performance(model)
}

#> Warning: There were 1 divergent transitions after warmup. Increasing adapt_delta above 0.95 may help. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is 1.05, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> # Indices of model performance
#> 
#>     ELPD ELPD_SE   LOOIC LOOIC_SE   WAIC   R2 R2_marginal R2_adjusted  RMSE
#>  -79.247   4.537 158.494    9.073 158.35 0.82       0.818       0.784 2.449