performance-package {performance}	R Documentation

performance: An R Package for Assessment, Comparison and Testing of Statistical Models

Description

A crucial aspect when building regression models is to evaluate the quality of modelfit. It is important to investigate how well models fit to the data and which fit indices to report. Functions to create diagnostic plots or to compute fit measures do exist, however, mostly spread over different packages. There is no unique and consistent approach to assess the model quality for different kind of models.

The primary goal of the performance package is to fill this gap and to provide utilities for computing indices of model quality and goodness of fit. These include measures like r-squared (R2), root mean squared error (RMSE) or intraclass correlation coefficient (ICC), but also functions to check (mixed) models for overdispersion, zero-inflation, convergence or singularity.

References: Lüdecke et al. (2021) doi: 10.21105/joss.03139

Details

performance-package

Author(s)

Maintainer: Daniel Lüdecke d.luedecke@uke.de (ORCID) (@strengejacke)

Authors:

Dominique Makowski dom.makowski@gmail.com (ORCID) (@Dom_Makowski) [contributor]

Mattan S. Ben-Shachar matanshm@post.bgu.ac.il (ORCID) (@mattansb) [contributor]

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets) [contributor]

Philip Waggoner philip.waggoner@gmail.com (ORCID) [contributor]

Brenton M. Wiernik brenton@wiernik.org (ORCID) (@bmwiernik) [contributor]

Rémi Thériault remi.theriault@mail.mcgill.ca (ORCID) (@rempsyc) [contributor]

Other contributors:

Vincent Arel-Bundock vincent.arel-bundock@umontreal.ca (ORCID) [contributor]

Martin Jullum [reviewer]

gjo11 [reviewer]

Etienne Bacher etienne.bacher@protonmail.com (ORCID) [contributor]

See Also

Useful links:

https://easystats.github.io/performance/

Report bugs at https://github.com/easystats/performance/issues

[Package performance version 0.11.0 Index]