Simple ROC curve

This function calculates a simple ROC curves of x/y coordinates based on response and predictions of a binomial model.

performance_roc(x, ..., predictions, new_data)

Arguments

x
A numeric vector, representing the outcome (0/1), or a model with binomial outcome.

...
One or more models with binomial outcome. In this case, new_data is ignored.

predictions
If x is numeric, a numeric vector of same length as x, representing the actual predicted values.

new_data
If x is a model, a data frame that is passed to predict() as newdata-argument. If NULL, the ROC for the full model is calculated.

Value

A data frame with three columns, the x/y-coordinate pairs for the ROC curve (Sensivity and Specifity), and a column with the model name.

Note

There is also a plot()-method implemented in the see-package.

Examples

library(bayestestR)

#> Note: The default CI width (currently `ci=0.89`) might change in future versions (see https://github.com/easystats/bayestestR/issues/250). To prevent any issues, please set it explicitly when using bayestestR functions, via the 'ci' argument.

data(iris)

set.seed(123)
iris$y <- rbinom(nrow(iris), size = 1, .3)
folds <- sample(nrow(iris), size = nrow(iris) / 8, replace = FALSE)
test_data <- iris[folds, ]
train_data <- iris[-folds, ]

model <- glm(y ~ Sepal.Length + Sepal.Width, data = train_data, family = "binomial")
performance_roc(model, new_data = test_data)

#> AUC: 37.66%

roc <- performance_roc(model, new_data = test_data)
area_under_curve(roc$Specifity, roc$Sensivity)

#> [1] 0.3766234

m1 <- glm(y ~ Sepal.Length + Sepal.Width, data = iris, family = "binomial")
m2 <- glm(y ~ Sepal.Length + Petal.Width, data = iris, family = "binomial")
m3 <- glm(y ~ Sepal.Length + Species, data = iris, family = "binomial")
performance_roc(m1, m2, m3)

#> # Area under Curve
#> 
#>   m1: 54.80%
#>   m2: 53.41%
#>   m3: 53.58%